Move test for store_bt_results to optimize_reports
